+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Changed
+- No default filename or special prefix
+
+### Fixed
+- Update dependancies
 
 ## [0.3.0][] - 2018-06-16
 ### Changed